pyproject.toml
[project]
name = "toshi-hazard-store"
version = "1.4.0"

homepage = "https://github.com/GNS-Science/toshi-hazard-store"
description = "Library for saving and retrieving NZHSM openquake hazard results with convenience (uses pyarrow with parquet format)."
authors = [
    { name = "Chris B Chamberlain", email = "chrisbc@artisan.co.nz" },
    {name="NSHM Compute Working Group", email="nshm@gns.cri.nz"}    
]
readme="README.md"
license = "AGPL-3.0-or-later"
classifiers=[
    'Development Status :: 2 - Pre-Alpha',
    'Intended Audience :: Developers',
    'Natural Language :: English',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
]
requires-python = ">=3.10,<4.0"
packages = [
    { include = "toshi_hazard_store" },
    { include = "tests", format = "sdist" },
]
dependencies = [
    "geopandas (>=1.1.2)",
    "lancedb (>=0.27.1)",
    "matplotlib (>=3.10.8)",
    "nshm-toshi-client (>=1.0.1)",
    "numpy (>=1.24)",
    "nzshm-common (>=0.8.4)",
    "nzshm-model (>=0.13.6)",
    "pandas (>=2,<3)",
    "pathy (>=0.11.0)",
    "pyarrow (>=20.0.0)",
    "pydantic (>=2.11.5)",
    "pynamodb (>=6.0.0)",
    "pynamodb-attributes (>=0.4.0)",
    "python-dotenv (>=1.0.1)",
    "s3path (>=0.6.1)",
    "toml (>=0.10.2)",
]

[project.scripts]
ths_compat = 'toshi_hazard_store.scripts.ths_compat:main'
ths_ds_check = 'toshi_hazard_store.scripts.ths_ds_check:main'
ths_ds_defrag = 'toshi_hazard_store.scripts.ths_ds_defrag:main'
ths_ds_filter = 'toshi_hazard_store.scripts.ths_ds_filter:main'
ths_grid_build = 'toshi_hazard_store.scripts.ths_grid_build:main'
ths_grid_sanity = 'toshi_hazard_store.scripts.ths_grid_sanity:main'
ths_rlz_import = 'toshi_hazard_store.scripts.ths_rlz_import:main'
ths_rlz_sanity = 'toshi_hazard_store.scripts.ths_rlz_sanity:main'

[dependency-groups]
openquake = [
    "fiona (>=1.9.5)",
    "networkx (>=3.2.1)",
    "numba",
    "pydantic-to-pyarrow (>=0.1.6)",
    "openquake-engine (>=3.20.1)"
]
scripts = [
    "boto3",
    "click (>=8.1.3)",
    "click-plugins (>=1.1.1)",
]
doc = [
    { include-group = "scripts" },
    "mkdocs (>=1.6.1)",
    "mkdocs-autorefs (>=1.0.1)",
    "mkdocs-click (>=0.8.1)",
    "mkdocs-include-markdown-plugin (>=7.1.8)",
    "mkdocs-material (>=9.5.21)",
    "mkdocs-material-extensions (>=1.3.1)",
    "mkdocs-pymdownx-material-extras (>=2.5.6)",
    "mkdocs",
    "mkdocstrings (>=0.26.2)",
    "mkdocstrings-python (>=1.18.2,<2.0.0)",
]

dev = [
    "black (>=24.10.0)",
    "bump2version (>=1.0.1)",
    "flake8 (>=7.1.1)",
    "flake8-docstrings (>=1.7.0)",
    "ipython (>=8.11.0)",
    "isort (>=5.13.2)",
    "matplotlib (>=3.6.3)",
    "moto (>=4)",
    "mypy (>=1.18.2)",
    "pandas-stubs (==2.3.0.250703.*)",
    "pip-audit (>=2.9.0)",
    "pytest-cov  (>=2.12.0)",
    "pytest-lazy-fixtures",
    "pytest-xdist (>=3.5.0)",
    "pyvista (>=0.44.1)",
    "safety (>=3.6.2)",
    "semver",
    "tox (>=4.4.4)",
    "twine (>=6.2.0,<7.0.0)",
    "types-requests (>=2.32.4.20250913,<3.0.0.0)",
    "types-toml (>=0.10.8.20240310,<0.11.0.0)",
    "pytest-mock (>=3.15.1,<4.0.0)",
]

[tool.black]
line-length = 120
skip-string-normalization = true
target-version = ['py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
skip_gitignore = true
# you can skip files as below
#skip_glob = docs/conf.py

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
